requirements.txt
streamlit
keras
tensorflow-cpu
pandas
matplotlib
plotly-express
networkx
Pillow
wordcloud